Move client api to backend.py
